refactor(useDisplayedFolder): Convert into Typescript
